Split CLI sketchlist from GUI one in prep for macros
